Moved VerifyPluginWorkAction and VerifyPluginWorkParameters to separate file VerifyPluginWorkAction.kt
