+++ src/main/kotlin/org/jetbrains/intellij/platform/gradle/tasks/VerifyPluginTask.kt
@@ -2,14 +2,12 @@
 
 package org.jetbrains.intellij.platform.gradle.tasks
 
-import org.apache.tools.ant.util.TeeOutputStream
 import org.gradle.api.DefaultTask
 import org.gradle.api.GradleException
 import org.gradle.api.Project
 import org.gradle.api.file.ConfigurableFileCollection
 import org.gradle.api.file.DirectoryProperty
 import org.gradle.api.file.RegularFileProperty
-import org.gradle.api.model.ObjectFactory
 import org.gradle.api.provider.ListProperty
 import org.gradle.api.provider.Property
 import org.gradle.api.tasks.*
@@ -17,22 +15,17 @@ import org.gradle.api.tasks.Optional
 import org.gradle.kotlin.dsl.assign
 import org.gradle.kotlin.dsl.get
 import org.gradle.kotlin.dsl.named
-import org.gradle.process.ExecOperations
-import org.gradle.workers.WorkAction
-import org.gradle.workers.WorkParameters
 import org.gradle.workers.WorkerExecutor
 import org.jetbrains.intellij.platform.gradle.Constants.Configurations
 import org.jetbrains.intellij.platform.gradle.Constants.Configurations.Attributes
 import org.jetbrains.intellij.platform.gradle.Constants.Plugin
 import org.jetbrains.intellij.platform.gradle.Constants.Tasks
 import org.jetbrains.intellij.platform.gradle.extensions.IntelliJPlatformExtension
-import org.jetbrains.intellij.platform.gradle.tasks.VerifyPluginTask.FailureLevel
 import org.jetbrains.intellij.platform.gradle.tasks.aware.PluginVerifierAware
 import org.jetbrains.intellij.platform.gradle.tasks.aware.RuntimeAware
 import org.jetbrains.intellij.platform.gradle.utils.Logger
 import org.jetbrains.intellij.platform.gradle.utils.asPath
 import org.jetbrains.intellij.platform.gradle.utils.extensionProvider
-import java.io.ByteArrayOutputStream
 import java.util.*
 import javax.inject.Inject
 import kotlin.io.path.exists
@@ -411,69 +404,3 @@ abstract class VerifyPluginTask : DefaultTask(), RuntimeAware, PluginVerifierAwa
         override fun toString() = name.lowercase().replace('_', '-')
     }
 }
-
-
-internal abstract class VerifyPluginWorkAction @Inject constructor(
-    private val execOperations: ExecOperations,
-    private val objectFactory: ObjectFactory
-) :  WorkAction<VerifyPluginWorkParameters> {
-
-    private val log = Logger(javaClass)
-
-    override fun execute() {
-
-        ByteArrayOutputStream().use { os ->
-            val outputStream = TeeOutputStream(System.out, os)
-
-            execOperations.javaexec {
-                mainClass.set("com.jetbrains.pluginverifier.PluginVerifierMain")
-
-                classpath = objectFactory.fileCollection().from(
-                    parameters.getPluginVerifierPath.get()
-                )
-
-                standardOutput = outputStream
-
-                args(parameters.getArgs.get())
-            }
-
-            verifyOutput(os.toString(), parameters.getFailureLevel)
-        }
-    }
-
-    /**
-     * @throws GradleException
-     */
-    @Throws(GradleException::class)
-    private fun verifyOutput(output: String, failureLevel: ListProperty<FailureLevel>, ) {
-        log.debug("Current failure levels: ${FailureLevel.values().joinToString(", ")}")
-
-        val invalidFilesMessage = "The following files specified for the verification are not valid plugins:"
-        if (output.contains(invalidFilesMessage)) {
-            val errorMessage = output.lines()
-                .dropWhile { it != invalidFilesMessage }
-                .dropLastWhile { !it.startsWith(" ") }
-                .joinToString("\n")
-
-            throw GradleException(errorMessage)
-        }
-
-        FailureLevel.values().forEach { level ->
-            if (failureLevel.get().contains(level) && output.contains(level.sectionHeading)) {
-                log.debug("Failing task on '$failureLevel' failure level")
-                throw GradleException(
-                    "$level: ${level.message} Check Plugin Verifier report for more details.\n" +
-                            "Incompatible API Changes: https://jb.gg/intellij-api-changes"
-                )
-            }
-        }
-    }
-}
-
-internal interface VerifyPluginWorkParameters : WorkParameters {
-    val getArgs: ListProperty<String>
-    val getPluginVerifierPath: Property<String>
-    val getFailureLevel: ListProperty<FailureLevel>
-}
-
-
